Fix broken escape sequences on FAQ and install pages
For the most part, text surrounded by backticks is automatically escaped, so the backslashes or HTML escapes were showing up on the page, which is confusing. For the FAQ page, I've removed the backslashes. For the install page, I've used the var element which is more semantically correct (and results in italic formatting); that's not directly exposed in markdown, so I had to use code tags instead of backticks.